+++ CHANGELOG.md
@@ -96,7 +96,7 @@ By @andyleiserson in [#9321](https://github.com/gfx-rs/wgpu/pull/9321).
 - Make `wgpu_types::texture::format::TextureChannel` accessible as `wgpu::TextureChannel`. By @TornaxO7 in [#9394](https://github.com/gfx-rs/wgpu/pull/9349).
 - Add support for `per_vertex` in Metal and DX12, as well as some validation for `per_vertex`, and a new enable extension, `wgpu_per_vertex`. By @inner-daemons in [#9219](https://github.com/gfx-rs/wgpu/pull/9219).
 - Add `ComputePass` version of `CommandEncoder::transition_resources` that allows intra-pass transitions. By @wingertge in [#9371](https://github.com/gfx-rs/wgpu/pull/9371).
-- Buffer slices with a length of 0 can now be created. By @beholdnec in [#8505] (https://github.com/gfx-rs/wgpu/pull/8505).
+- Buffer slices with a length of 0 can now be instantiated. By @beholdnec in [#8505] (https://github.com/gfx-rs/wgpu/pull/8505).
 
 #### Metal